requirements.txt

python-binance
twilio
requests
